Add caption to map, update map instructions
Added caption to map explaining placeholder (pink markers); Added Placeholder field to Data Dictionary. Updated map instructions to emphasize providing .csv file with correct lat/ long to Sunshine Request and importance of complete addresses to getting lat/ long.

I believe that this is the final update to the website text before we are ready to transfer to Sunshine Request.